style(pacquet): satisfy perfectionist single-letter and layout lints
- Rename the reporter generic `R` to `Reporter` (bounded `self::Reporter`)
  in pack's `api` / `run_scripts_if_present` and the CLI `pack` command,
  matching the convention already used in executor, git-fetcher, and
  package-manager (perfectionist::single_letter_generic).
- Rename `locale_compare_en`'s `a` / `b` parameters to `left` / `right`
  (perfectionist::single_letter_function_param).
- Move the inline `transform` test module into an external
  `transform/tests.rs` (perfectionist::unit_test_file_layout).
